20240912_김광훈_관리자 메뉴 수정

+++ ECR/src/main/java/com/project/springboot/repository/MemberRepository.java
@@ -13,7 +13,7 @@ public interface MemberRepository extends JpaRepository<Member, String> {
 
 	Optional<Member> findMemberByLoginTypeAndMemberId(int loginType, String memberId);
 	
-	List<Member> findByLoginType(int loginType);
+	List<Member> findByLoginType(Long loginType);
 
 }
